requirements.txt
discord.py
python-dotenv
requests
lxml
pytz
discord-py-slash-command
apscheduler